Added common c and h and in progress on table refactor of c_nn_v2_fast.c
