+++ plaster/run/nn_v2/fast/c_nn_v2_fast.pxd
@@ -1,13 +1,26 @@
 cdef extern from "c_nn_v2_fast.h":
+    ctypedef unsigned char Uint8
     ctypedef unsigned long Uint64
     ctypedef unsigned long Size
-    ctypedef unsigned int Size32
     ctypedef unsigned long Index
-    ctypedef unsigned int Index32
+    ctypedef unsigned long HashKey
     ctypedef unsigned char DyeType
+    ctypedef unsigned char CycleKindType
+    ctypedef unsigned long PIType
+    ctypedef double RecallType
+    ctypedef unsigned int Size32
+    ctypedef unsigned int Index32
     ctypedef float RadType
     ctypedef float Score
     ctypedef float WeightType
+
+    ctypedef struct Table:
+        Uint8 *rows
+        Uint64 n_bytes_per_row
+        Uint64 n_max_rows
+        Uint64 n_rows
+
+    Table table_init(Uint8 *base, Size n_bytes, Size n_bytes_per_row)
 
     ctypedef struct DyePepRec:
         Size count
@@ -18,15 +31,11 @@ cdef extern from "c_nn_v2_fast.h":
         Size n_neighbors
         Size n_cols
 
-        Size test_unit_radmat_n_rows
-        RadType *test_unit_radmat
-
-        Size train_dyemat_n_rows
-        RadType *train_dyemat
-        WeightType *train_dyetrack_weights
-
-        Index32 *output_pred_iz
-        Score *output_scores
+        Table test_unit_radmat
+        Table train_dyetrack_weights
+        Table train_dyemat
+        Table output_pred_iz
+        Table output_scores
 
     void context_start(Context *ctx)
     void context_free(Context *ctx)
