+++ plaster/run/sim_v2/fast/c_sim_v2_fast.h
@@ -1,31 +1,7 @@
 #ifndef SIM_V2_H
 #define SIM_V2_H
 
-// See sim.c for docs.
-
-typedef __uint8_t Uint8;
-typedef __uint64_t Uint64;
-typedef __uint128_t Uint128;
-typedef __int128_t Sint128;
-typedef __int64_t Sint64;
-typedef double Float64;
-
-
-typedef Uint64 Size;
-typedef Uint64 Index;
-typedef Uint64 HashKey;
-typedef Uint8 DyeType;
-typedef Uint8 CycleKindType;
-typedef Uint64 PIType;
-typedef Float64 RecallType;
-
-
-#define N_MAX_CHANNELS ((DyeType)(8))
-#define NO_LABEL ((DyeType)(N_MAX_CHANNELS - 1))
-#define N_MAX_CYCLES ((DyeType)64)
-#define CYCLE_TYPE_PRE ((CycleKindType)(0))
-#define CYCLE_TYPE_MOCK ((CycleKindType)(1))
-#define CYCLE_TYPE_EDMAN ((CycleKindType)(2))
+#include "c_common.h"
 
 
 typedef struct {
@@ -42,27 +18,12 @@ typedef struct {
 
 
 typedef struct {
-    Uint8 *rows;
-    Uint64 n_bytes_per_row;
-    Uint64 n_max_rows;
-    Uint64 n_rows;
-} Table;
-
-
-typedef struct {
     Size count;
     Index dtr_i;
     DyeType chcy_dye_counts[];
     // Note, this is a variable sized record
     // See dtr_* functions for manipulating it
 } DTR;  // DTR = Dye Track Record
-
-
-typedef struct {
-    Size count;
-    Index dtr_i;
-    Index pep_i;
-} DyePepRec;
 
 
 typedef struct {
@@ -94,7 +55,6 @@ int setup_sanity_checks(Size n_channels, Size n_cycles);
 Uint64 prob_to_p_i(double p);
 void rand64_seed(Uint64 seed);
 Size dtr_n_bytes(Size n_channels, Size n_cycles);
-Table table_init(Uint8 *base, Size n_bytes, Size n_bytes_per_row);
 Hash hash_init(HashRec *buffer, Size n_max_recs);
 void context_work_orders_start(Context *ctx);
 Index context_dtr_get_count(Context *ctx, Index dtr_i);
